backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "kokoro-onnx>=0.4.9",
    "numpy>=2.4.1",
    "pydantic-settings>=2.12.0",
    "soundfile>=0.13.1",
    "uvicorn>=0.40.0",
]
